Remove websoc-api typescript declarations
